{{ project_name }}/requirements.txt
Django==1.11.2
django-environ==0.4.3
Jinja2==2.9.6
pytz==2017.2
six==1.10.0
whitenoise==3.3.0
